requirements.txt
neo4j==5.9.0
openai
pinecone-client==2.2.0
pyvis==0.3.1
networkx==3.1
tqdm
python-dotenv
requests
huggingface_hub
aiohttp==3.9.5
sentence-transformers
